Drop default umask of 0077
For most backends, it doesn't actually take effect. And it can be
confusing for people that back up to a drive that is ext4-formatted
but then try to restore on a new system.

If folks are worried about others accessing the backup files,
encryption is the recommended path for that.

https://gitlab.com/duplicity/duplicity/-/issues/24