Report errors if B2 backend does exist but otherwise fails to import
Sometimes import can fail because one of B2's dependencies is broken.

The trick here is to query the "name" attribute of ModuleNotFoundError
to see if B2 is the module that failed. Unfortunately this only works on
Python 3.6+. In older versions, the original behavior is retained.

This partially mitigates the issue described in
https://github.com/henrysher/duplicity/issues/14.