List required volumes when called with 'restore --dry-run'
When restoring in dry-run mode, and with the manifest available, list
the volumes that would be gotten from the backend when actually
performing the operation.
This is intended to aid users of e.g. the S3 backend with (deep) glacier
storage, allowing the following workflow to recover files, optionally at
a certain time, from a long-term archive:
1. duplicity restore --dry-run [--file-to-restore <file/dir>] [--time <time>] boto3+s3://...
2. Start a Glacier restore process for all the listed volumes
3. duplicity restore [--file-to-restore <file/dir>] [--time <time>] boto3+s3://...

Signed-off-by: Matthias Blankertz <matthias@blankertz.org>