Implement megav3 backend to to cater for change in MEGACmd
MEGACmd no longer having seperate mega-cmd and mega-cmd-server, there is
only mega-cmd, and when we call any mega-xxx command, mega-cmd will be
started in the background first.

megav3 is implemented to cater this change, it is a fork of
megav2backend but with modification to support the new MEGACmd usage.

It also support a new query param in the url "no_logout" to tell the
duplicity not to logout from current mega session.