Implement Box backend
This backend dependes on boxsdk[jwt].
It also assume user have created a custom app on Box.com and have exported
a config.json.

The URL for this backend is as follow:
box://{PATH_TO_DIR}[?config={PATH_TO_CONFIG_FILE}]

The `config` query param could be omitted if BOX_CONFIG_PATH envvar is
defined and set to PATH_TO_CONFIG_FILE. When both of them is defined,
query param takes precedence.

A short summary on configurating the Box custom App are:
1. Choose "App Access Only" for "App Access Level"
2. Check "Write all files and folders stored in Box"
3. Generate a Public/Private Keypair

The user also need to grant the created custom app permission in the
admin console (https://app.box.com/master/custom-apps) by clicking
the "+" button and enter the client_id which can be found on the custom
app's configuration page.